+++ hspec-core/src/Test/Hspec/Core/Example/Location.hs
@@ -4,6 +4,7 @@ module Test.Hspec.Core.Example.Location (
 , extractLocation
 
 -- for testing
+, parseAssertionFailed
 , parseCallStack
 , parseLocation
 , parseSourceSpan
@@ -31,6 +32,21 @@ extractLocation e =
   <|> locationFromPatternMatchFail e
   <|> locationFromRecConError e
   <|> locationFromIOException e
+  <|> locationFromNoMethodError e
+  <|> locationFromAssertionFailed e
+
+locationFromNoMethodError :: SomeException -> Maybe Location
+locationFromNoMethodError e = case fromException e of
+  Just (NoMethodError s) -> listToMaybe (words s) >>= parseSourceSpan
+  Nothing -> Nothing
+
+locationFromAssertionFailed :: SomeException -> Maybe Location
+locationFromAssertionFailed e = case fromException e of
+  Just (AssertionFailed loc) -> parseAssertionFailed loc
+  Nothing -> Nothing
+
+parseAssertionFailed :: String -> Maybe Location
+parseAssertionFailed loc = parseCallStack loc <|> parseSourceSpan loc
 
 locationFromErrorCall :: SomeException -> Maybe Location
 locationFromErrorCall e = case fromException e of